Update _dropdown.scss
Fixes #49 .

As the new z-index default in Bootstrap5 for a modal is 1055 (instead of 1050), the z-index for the dropdown should be 1056 (instead of 1051). Currently this default 1051 value is handled in the select2 main repository (https://github.com/select2/select2/blob/develop/src/scss/_dropdown.scss), which shouldn't change to stay compatible with Bootstrap4. Hence that PR in this project.